refactor(plugin-flow-builder): organize files for manage user-input outside action folder
