+++ packages/botonic-plugin-flow-builder/src/user-input/intent.ts
@@ -2,7 +2,7 @@ import { ActionRequest } from '@botonic/react'
 
 import { FlowBuilderApi } from '../api'
 import { HtIntentNode } from '../content-fields/hubtype-fields'
-import { EventName, trackEvent } from './tracking'
+import { EventName, trackEvent } from '../tracking'
 
 export async function getIntentNodeByInput(
   cmsApi: FlowBuilderApi,
